Implement fsm from assignment 3 into worker agent, since only worker agent seems have more than 1 purpose

+++ src/Timer.h
@@ -17,7 +17,7 @@ struct Timer {
     bool timerFinished = false;
     float timeCap = 1;
 
-    void setNewTimer(int _newCap) {
+    void setNewTimer(float _newCap) {
         timerFinished = false;
         timeCap = _newCap;
         currTimePassed = 0;
